refactor(daemon): extract daemon HTTP request guards + live-artifact serving helpers
Strangler-fig slice 3: move the daemon's request/response-edge security and
serving helpers out of server.ts into a new sibling module, byte-identical.

Four related sub-groups:
- Local-daemon request validation (loopback peer/host/origin) + the
  requireLocalDaemonRequest middleware.
- Project preview-scope registry (mint/validate short-lived preview scopes) +
  preview asset-path parsing.
- Tool-request authorization (bearer token -> toolTokenRegistry grant) +
  project/run override checks.
- Live-artifact route serving (archive filename sanitize, route-error mapping,
  preview/code response headers).

All deps are imports (net, randomUUID, toolTokenRegistry, sendApiError, the
live-artifact error classes) -- no daemon module state, no cycle. server.ts
imports back the 13 symbols it references; the 6 internal-only helpers stay
module-private. The four sub-groups are a candidate for a later split.

No behavior change. Orphan/duplicate-binding checks clean.
