+++ apps/daemon/src/daemon-request-guards.ts
@@ -1,0 +1,289 @@
+// Authors: Leon Aburime using Claude Fable 5
+// @ts-nocheck — carried over verbatim from server.ts's file-level @ts-nocheck.
+// The moved bodies are untyped JS-in-TS; typing them is a later effort and new
+// sibling code must NOT copy this.
+/** @module daemon-request-guards
+ * Security and serving helpers for the daemon's HTTP routes. Four related
+ * concerns that all live at the request/response edge:
+ *
+ * - Local-daemon request validation: `normalizeLocalAuthority`,
+ *   `isLoopbackHostname`, `isLoopbackPeerAddress`, `localOriginFromHeader`,
+ *   `validateLocalDaemonRequest`, `requireLocalDaemonRequest` (the express
+ *   middleware) reject any request that is not a loopback peer/host/origin.
+ * - Project preview-scope tracking: `createProjectPreviewScopeRegistry` mints
+ *   and validates short-lived per-project preview scopes; `parseProjectPreviewAssetPath`
+ *   parses the preview asset URL shape.
+ * - Tool-request authorization: `bearerTokenFromRequest`, `authorizeToolRequest`,
+ *   `optionalToolGrantFromRequest`, `requestProjectOverride`, `requestRunOverride`
+ *   validate the tool token grant against the shared toolTokenRegistry.
+ * - Live-artifact route serving: `sanitizeArchiveFilename`,
+ *   `sendLiveArtifactRouteError`, `setLiveArtifactPreviewHeaders`,
+ *   `setLiveArtifactCodeHeaders`.
+ *
+ * All depend only on imported services (toolTokenRegistry, sendApiError, the
+ * live-artifact error classes), never on daemon module-level mutable state.
+ *
+ * Extracted verbatim from apps/daemon/src/server.ts (strangler-fig slice 3);
+ * server.ts imports back the thirteen symbols it references. The four sub-groups
+ * are a candidate for a later split into cohesive modules.
+ */
+
+import net from 'node:net';
+import { randomUUID } from 'node:crypto';
+import { toolTokenRegistry } from './tool-tokens.js';
+import { sendApiError } from './http/api-errors.js';
+import { ConnectorServiceError } from './connectors/service.js';
+import {
+  LiveArtifactRefreshLockError,
+  LiveArtifactStoreValidationError,
+} from './live-artifacts/store.js';
+import { LiveArtifactRefreshUnavailableError } from './live-artifacts/refresh-service.js';
+import { LiveArtifactRefreshAbortError } from './live-artifacts/refresh.js';
+
+// Filename slug for the Content-Disposition header on archive downloads.
+// Browsers reject quotes and control bytes; we keep Unicode letters/digits
+// so a project name with non-ASCII characters (e.g. "café-design")
+// survives instead of becoming a row of underscores.
+export function sanitizeArchiveFilename(raw) {
+  const cleaned = String(raw ?? '')
+    .replace(/[\\/:*?"<>|]/g, '_')
+    .replace(/[\u0000-\u001f\u007f]/g, '')
+    .replace(/\s+/g, '-')
+    .replace(/^-+|-+$/g, '')
+    .slice(0, 80);
+  return cleaned;
+}
+
+export function sendLiveArtifactRouteError(res, err) {
+  if (err instanceof LiveArtifactStoreValidationError) {
+    return sendApiError(res, 400, 'LIVE_ARTIFACT_INVALID', err.message, {
+      details: { kind: 'validation', issues: err.issues },
+    });
+  }
+  if (err instanceof LiveArtifactRefreshLockError) {
+    return sendApiError(res, 409, 'REFRESH_LOCKED', err.message, {
+      details: { artifactId: err.artifactId },
+    });
+  }
+  if (err instanceof LiveArtifactRefreshUnavailableError) {
+    return sendApiError(res, 400, 'LIVE_ARTIFACT_REFRESH_UNAVAILABLE', err.message);
+  }
+  if (err instanceof LiveArtifactRefreshAbortError) {
+    return sendApiError(res, err.kind === 'cancelled' ? 499 : 504, 'LIVE_ARTIFACT_REFRESH_TIMEOUT', err.message, {
+      details: { kind: err.kind, timeoutMs: err.timeoutMs ?? null, step: err.step ?? null },
+    });
+  }
+  if (err instanceof ConnectorServiceError) {
+    return sendApiError(res, err.status, err.code, err.message, err.details === undefined ? {} : { details: err.details });
+  }
+  if (err && typeof err === 'object' && 'code' in err && err.code === 'ENOENT') {
+    return sendApiError(res, 404, 'LIVE_ARTIFACT_NOT_FOUND', 'live artifact not found');
+  }
+  return sendApiError(res, 500, 'LIVE_ARTIFACT_STORAGE_FAILED', String(err));
+}
+
+function normalizeLocalAuthority(value) {
+  if (typeof value !== 'string') return null;
+  const trimmed = value.trim();
+  if (!trimmed || /[\s/@]/.test(trimmed) || trimmed.includes(',')) return null;
+
+  try {
+    const parsed = new URL(`http://${trimmed}`);
+    const hostname = parsed.hostname.toLowerCase().replace(/\.$/, '');
+    if (!hostname || parsed.username || parsed.password || parsed.pathname !== '/') return null;
+    return { hostname, port: parsed.port };
+  } catch {
+    return null;
+  }
+}
+
+export function isLoopbackHostname(hostname) {
+  const normalized = String(hostname || '').toLowerCase().replace(/^\[|\]$/g, '').replace(/\.$/, '');
+  if (normalized === 'localhost') return true;
+  if (normalized === '::1' || normalized === '0:0:0:0:0:0:0:1') return true;
+  if (net.isIP(normalized) === 4) return normalized === '127.0.0.1' || normalized.startsWith('127.');
+  return false;
+}
+
+export function isLoopbackPeerAddress(address) {
+  if (typeof address !== 'string') return false;
+  const normalized = address.trim().toLowerCase().replace(/^\[|\]$/g, '');
+  if (!normalized) return false;
+  if (normalized.startsWith('::ffff:')) return isLoopbackPeerAddress(normalized.slice('::ffff:'.length));
+  if (normalized === '::1' || normalized === '0:0:0:0:0:0:0:1') return true;
+  if (net.isIP(normalized) === 4) return normalized === '127.0.0.1' || normalized.startsWith('127.');
+  return false;
+}
+
+const PROJECT_PREVIEW_SCOPE_TTL_MS = 60 * 60 * 1000;
+const PROJECT_PREVIEW_ASSET_PATH_RE = /^\/projects\/([^/]+)\/preview\/([^/]+)\/.+$/u;
+
+export function createProjectPreviewScopeRegistry() {
+  const scopes = new Map();
+
+  function pruneExpired(now = Date.now()) {
+    for (const [scope, entry] of scopes) {
+      if (entry.expiresAt <= now) scopes.delete(scope);
+    }
+  }
+
+  return {
+    mint(projectId) {
+      pruneExpired();
+      const scope = randomUUID();
+      scopes.set(scope, {
+        projectId: String(projectId),
+        expiresAt: Date.now() + PROJECT_PREVIEW_SCOPE_TTL_MS,
+      });
+      return scope;
+    },
+    validate(projectId, scope) {
+      const key = String(scope || '');
+      const entry = scopes.get(key);
+      if (!entry) return false;
+      if (entry.expiresAt <= Date.now()) {
+        scopes.delete(key);
+        return false;
+      }
+      return entry.projectId === String(projectId);
+    },
+  };
+}
+
+export function parseProjectPreviewAssetPath(pathname) {
+  const match = PROJECT_PREVIEW_ASSET_PATH_RE.exec(String(pathname || ''));
+  if (!match) return null;
+  try {
+    return {
+      projectId: decodeURIComponent(match[1]),
+      scope: match[2],
+    };
+  } catch {
+    return null;
+  }
+}
+
+function localOriginFromHeader(value) {
+  if (typeof value !== 'string') return null;
+  const trimmed = value.trim();
+  if (!trimmed || trimmed === 'null' || trimmed.includes(',')) return null;
+
+  try {
+    const parsed = new URL(trimmed);
+    if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:') return null;
+    if (parsed.pathname !== '/' || parsed.search || parsed.hash || parsed.username || parsed.password) return null;
+    if (!isLoopbackHostname(parsed.hostname)) return null;
+    return parsed.origin;
+  } catch {
+    return null;
+  }
+}
+
+function validateLocalDaemonRequest(req) {
+  if (!isLoopbackPeerAddress(req.socket?.remoteAddress)) {
+    return {
+      ok: false,
+      message: 'request peer must be a loopback address',
+      details: { peer: 'remoteAddress' },
+    };
+  }
+
+  const host = normalizeLocalAuthority(req.get('host'));
+  if (!host || !isLoopbackHostname(host.hostname)) {
+    return {
+      ok: false,
+      message: 'request host must be a loopback daemon address',
+      details: { header: 'host' },
+    };
+  }
+
+  const originHeader = req.get('origin');
+  if (originHeader !== undefined && !localOriginFromHeader(originHeader)) {
+    return {
+      ok: false,
+      message: 'request origin must be a loopback daemon origin',
+      details: { header: 'origin' },
+    };
+  }
+
+  return { ok: true, origin: localOriginFromHeader(originHeader) };
+}
+
+export function requireLocalDaemonRequest(req, res, next) {
+  const validation = validateLocalDaemonRequest(req);
+  if (!validation.ok) {
+    return sendApiError(res, 403, 'FORBIDDEN', validation.message, validation.details ? { details: validation.details } : {});
+  }
+
+  res.setHeader('Vary', 'Origin');
+  if (validation.origin) {
+    res.setHeader('Access-Control-Allow-Origin', validation.origin);
+  }
+  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS');
+  res.setHeader('Access-Control-Allow-Headers', 'Content-Type');
+  res.setHeader('Access-Control-Max-Age', '600');
+  next();
+}
+
+export function setLiveArtifactPreviewHeaders(res) {
+  res.setHeader('Content-Type', 'text/html; charset=utf-8');
+  res.setHeader('Cache-Control', 'no-store');
+  res.setHeader('X-Content-Type-Options', 'nosniff');
+  res.setHeader('Referrer-Policy', 'no-referrer');
+  res.setHeader(
+    'Content-Security-Policy',
+    [
+      "default-src 'none'",
+      "base-uri 'none'",
+      "script-src 'none'",
+      "object-src 'none'",
+      "connect-src 'none'",
+      "form-action 'none'",
+      "frame-ancestors 'self'",
+      "img-src 'self' data: blob:",
+      "font-src 'self' data:",
+      "style-src 'unsafe-inline'",
+      'sandbox allow-same-origin',
+    ].join('; '),
+  );
+}
+
+export function setLiveArtifactCodeHeaders(res) {
+  res.setHeader('Content-Type', 'text/plain; charset=utf-8');
+  res.setHeader('Cache-Control', 'no-store');
+  res.setHeader('X-Content-Type-Options', 'nosniff');
+  res.setHeader('Referrer-Policy', 'no-referrer');
+}
+
+function bearerTokenFromRequest(req) {
+  const header = req.get('authorization');
+  if (typeof header !== 'string') return undefined;
+  const match = /^Bearer\s+(.+)$/i.exec(header.trim());
+  return match?.[1];
+}
+
+export function authorizeToolRequest(req, res, operation) {
+  const endpoint = req.path;
+  const validation = toolTokenRegistry.validate(bearerTokenFromRequest(req), { endpoint, operation });
+  if (!validation.ok) {
+    const status = validation.code === 'TOOL_ENDPOINT_DENIED' || validation.code === 'TOOL_OPERATION_DENIED' ? 403 : 401;
+    sendApiError(res, status, validation.code, validation.message, {
+      details: { endpoint, operation },
+    });
+    return null;
+  }
+  return validation.grant;
+}
+
+export function optionalToolGrantFromRequest(req, options = {}) {
+  const validation = toolTokenRegistry.validate(bearerTokenFromRequest(req), options);
+  return validation.ok ? validation.grant : null;
+}
+
+export function requestProjectOverride(projectId, tokenProjectId) {
+  return typeof projectId === 'string' && projectId.length > 0 && projectId !== tokenProjectId;
+}
+
+export function requestRunOverride(runId, tokenRunId) {
+  return typeof runId === 'string' && runId.length > 0 && runId !== tokenRunId;
+}
